requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Werkzeug
requests
python-dotenv
gunicorn
